requirements.txt
pandas
yfinance
matplotlib
python-binance
